Split library between cmd line and functionality
